src/A. staging database/4. Local staging merge to central staging/requirements.txt
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30